minor refactor: restructure repo to adhere to tmpl
