+++ http.worker.dockerfile
@@ -1,18 +1,24 @@
 FROM python:3.10-slim
 RUN pip install -U pip
 
-RUN mkdir -p /siibra_jugex/siibra_jugex_http
-COPY ./siibra_jugex_http/requirements-worker.txt /siibra_jugex/siibra_jugex_http/requirements-worker.txt
-RUN pip install -r /siibra_jugex/siibra_jugex_http/requirements-worker.txt
+RUN mkdir /requirements
 
-COPY ./requirements.txt /siibra_jugex/requirements.txt
-RUN pip install -r /siibra_jugex/requirements.txt
+# install requirements for worker (celery etc)
+COPY ./http_wrapper/requirements-worker.txt /requirements/
+RUN pip install -r /requirements/requirements-worker.txt
 
-COPY . /siibra_jugex
-WORKDIR /siibra_jugex
+# install requirements for toolbox (numpy, etc)
+COPY ./requirements.txt /requirements/requirements.txt
+RUN pip install -r /requirements/requirements.txt
 
+# NB
+# the path is chosen deliberately to allow for proper config mapping
+# modify this at your own peril
+COPY . /siibra_toolbox
+WORKDIR /siibra_toolbox
 RUN pip install .
 
+WORKDIR /siibra_toolbox
 USER nobody
 
-ENTRYPOINT celery -A siibra_jugex_http.scheduling.worker.app worker -l INFO
+ENTRYPOINT celery -A http_wrapper.scheduling.worker.app worker -l INFO
