+++ http_wrapper/exceptions.py
