+++ http_wrapper/server.py
@@ -1,21 +1,12 @@
-from siibra_jugex_http.routes.analysis import router as analysis_router
-from siibra_jugex_http.routes.notebook import router as notebook_router
-from siibra_jugex_http.jugex_logger import access_logger
+from http_wrapper.routes.analysis import router as analysis_router
+from http_wrapper.routes.notebook import router as notebook_router
+from http_wrapper.jugex_logger import access_logger
 
 from fastapi import FastAPI, Request
 from fastapi.staticfiles import StaticFiles
 from fastapi.middleware.cors import CORSMiddleware
-from os import path, getenv
+from os import getenv
 import time
-
-path_to_static = path.join(
-    path.dirname(__file__),
-    "../siibra_jugex_viewerplugin",
-    "public/"
-)
-
-if getenv("SIIBRA_JUGEX_STATIC_DIR"):
-    path_to_static = getenv("SIIBRA_JUGEX_STATIC_DIR")
 
 app = FastAPI()
 
@@ -44,5 +35,8 @@ def hello():
 
 app.include_router(analysis_router, prefix="/analysis")
 app.include_router(notebook_router, prefix="/notebook")
-app.mount('/viewer_plugin', StaticFiles(directory=path_to_static))
+
+if getenv("SIIBRA_TOOLBOX_VIEWER_PLUGIN_STATIC_DIR"):
+    path_to_viewer_plugin = getenv("SIIBRA_TOOLBOX_VIEWER_PLUGIN_STATIC_DIR")
+    app.mount('/viewer_plugin', StaticFiles(directory=path_to_viewer_plugin))
