+++ viewer_plugin/package.json
