+++ viewer_plugin/public/manifest.json
