+++ viewer_plugin/scripts/setupTypeScript.js
