+++ viewer_plugin/src/App.svelte
