+++ viewer_plugin/src/JugexSlider.svelte
