+++ viewer_plugin/src/RoiSelection.svelte
