+++ viewer_plugin/src/ShowResult.svelte
