+++ viewer_plugin/src/store.js
