+++ viewer_plugin/src/theme/_smui-theme.scss
