+++ viewer_plugin/src/theme/dark/_smui-theme.scss
